refactor: move game state into seperate module
