+++ Makefile
@@ -1,6 +1,6 @@
 CC=gcc
 CFLAGS=-Wall --std=c99 
-OBJECTS=./build/vector.o ./build/my_string.o ./build/node.o ./build/word_linked_list.o ./build/game_print.o ./build/user_input.o ./build/dictionary.o
+OBJECTS=./build/vector.o ./build/my_string.o ./build/node.o ./build/word_linked_list.o ./build/game_state.o ./build/game_print.o ./build/user_input.o ./build/dictionary.o
 
 main: ./src/main.c $(OBJECTS)
 	$(CC) $(CFLAGS) ./src/main.c $(OBJECTS) -o play_evil_hangman
@@ -15,6 +15,9 @@ main: ./src/main.c $(OBJECTS)
 	$(CC) $(CFLAGS) -c -o $@ $<
 
 ./build/word_linked_list.o: ./src/data_types/word_linked_list/word_linked_list.c
+	$(CC) $(CFLAGS) -c -o $@ $<
+
+./build/game_state.o: ./src/game_state/game_state.c
 	$(CC) $(CFLAGS) -c -o $@ $<
 
 ./build/game_print.o: ./src/ui/game_print.c
